Separated request functionality into extension Added dynamic loading of extensions from 'extensions' dir removed old function from radarr_integration
